Задание 9.13.3
